extract report methods into their own controller and rename functions
