+++ grails-app/controllers/streama/VideoController.groovy
@@ -211,74 +211,4 @@ class VideoController {
     respond result
   }
 
-  def sendErrorReport() {
-    def jsonData = request.JSON
-    Video currentVideo = Video.get(jsonData.videoId)
-    String errorCode = jsonData.errorCode
-    User currentUser = springSecurityService.currentUser
-    Integer preExistingReportCount = Report.where {
-      and {
-        eq('video', currentVideo)
-        eq('createdBy', currentUser)
-        ne('resolved', true)
-        eq('errorCode', errorCode)
-      }
-    }.count()
-    if (preExistingReportCount > 0) {
-      render(status: 412, text: 'Report already sent.')
-      return
-    }
-    Report newReport = new Report()
-    newReport.errorCode = errorCode
-    newReport.createdBy = currentUser
-    newReport.video = currentVideo
-    newReport.resolved = false
-    newReport.save()
-    respond newReport
-  }
-
-  def getErrorReports () {
-    def reports  = Report.list()
-      return [reportList:reports]
-  }
-
-  @Transactional
-  def resolveReport() {
-    def jsonData = request.JSON
-    def report = Report.get(jsonData.reportId)
-    report.resolved = true
-    report.lastUpdated = new Date()
-    report.save()
-
-    respond report
-  }
-
-  @Transactional
-  def unresolveReport() {
-    def jsonData = request.JSON
-    def report = Report.get(jsonData.reportId)
-    report.resolved = false
-    report.lastUpdated = new Date()
-    report.save()
-
-    respond report
-  }
-
-  @Transactional
-  def resolveMultipleReports() {
-    def jsonData = request.JSON
-    def resolvedReports = []
-    jsonData.ids.each { id ->
-      def report = Report.get(id)
-      report.resolved = true
-      report.lastUpdated = new Date()
-      report.save()
-      resolvedReports.add([
-        id:id,
-        resolved:report.resolved,
-        lastUpdated:report.lastUpdated
-      ])
-    }
-    respond resolvedReports
-  }
 }
